HomeWork3 fixes according to review
