+++ src/entities/comment/types/types.ts
@@ -1,0 +1,6 @@
+export type PostCommentProps = {
+  id: string;
+  commentText: string;
+  author: string;
+  date: Date;
+};
