+++ src/entities/comment/ui/Comment.module.scss
@@ -1,0 +1,11 @@
+.commentsBlock__comment_other {
+  display: flex;
+  font-size: 12px;
+  justify-content: start;
+  gap:10px;
+}
+
+.commentsBlock__comment_text {
+  justify-self: start;
+  text-align: start;
+}
