+++ src/shared/ui/Modal/Modal.module.scss
@@ -27,16 +27,3 @@
     color: var(--text-color);
     border:1px solid var(--text-color);
 }
-
-.modalHeader {
-}
-
-.modalText {
-}
-
-.modalBody {
-    padding:20px;
-}
-
-.modalFooter {
-}
