+++ src/shared/ui/Modal/Modal.tsx
@@ -1,34 +1,8 @@
 import styles from './Modal.module.scss';
 import { createPortal } from 'react-dom';
-import {
-  createContext,
-  type PropsWithChildren,
-  type ReactNode,
-  useContext,
-} from 'react';
-import { Button } from '../Button/Button.tsx';
-
-type Modal = {
-  title: ReactNode | string;
-  body: ReactNode | string;
-  footer: ReactNode | string;
-};
-
-type ModalContext = {
-  modal: Modal;
-  handleClose: () => void;
-};
-
-const ModalContext = createContext<ModalContext | undefined>(undefined);
-
-function useModalContext() {
-  const context = useContext(ModalContext);
-  if (!context) {
-    throw new Error('useModalContext must be used within ModalContextProvider');
-  }
-  return context;
-}
-
+import { type PropsWithChildren } from 'react';
+import { type Modal, ModalContext } from '../../lib/modal/ModalContext.tsx';
+import { Title, Body, Footer, CloseButton } from './components';
 type ModalProps = PropsWithChildren & {
   modal: Modal;
   handleClose: () => void;
@@ -49,34 +23,7 @@ export function Modal({ handleClose, children, modal }: ModalProps) {
   );
 }
 
-Modal.Title = function ModalTitle() {
-  const { modal } = useModalContext();
-
-  return <h3 className={styles.modalHeader}>{modal.title}</h3>;
-};
-
-Modal.Body = function ModalBody() {
-  const { modal } = useModalContext();
-
-  if (typeof modal.body === 'string')
-    return <p className={styles.modalText}>{modal.body}</p>;
-  else {
-    return <div className={styles.modalBody}>{modal.body}</div>;
-  }
-};
-
-Modal.Footer = function ModalFooter() {
-  const { modal } = useModalContext();
-
-  return <p className={styles.modalFooter}>{modal.footer}</p>;
-};
-
-Modal.CloseButton = function ModalCloseButton() {
-  const { handleClose } = useModalContext();
-
-  return (
-    <Button buttonType={'closeModalCross'} onClick={handleClose}>
-      ✖
-    </Button>
-  );
-};
+Modal.Title = Title;
+Modal.Body = Body;
+Modal.Footer = Footer;
+Modal.CloseButton = CloseButton;
