+++ src/shared/ui/Modal/components/Body/Body.module.scss
@@ -1,0 +1,7 @@
+.modalBody {
+  padding:20px;
+}
+
+.modalText {
+}
+
