+++ src/shared/ui/Modal/components/Footer/Footer.module.scss
@@ -1,0 +1,2 @@
+.modalFooter {
+}
