+++ src/shared/ui/Modal/components/Title/Title.module.scss
@@ -1,0 +1,2 @@
+.modalHeader {
+}
