+++ src/shared/ui/Modal/components/index.ts
@@ -1,0 +1,4 @@
+export { Title } from './Title/Title.tsx';
+export { Body } from './Body/Body.tsx';
+export { Footer } from './Footer/Footer.tsx';
+export { CloseButton } from './CloseButton/CloseButton.tsx';
