+++ src/widgets/CommentList/CommentList.tsx
@@ -1,5 +1,6 @@
 import styles from './CommentList.module.scss';
 import type { PostProps } from '../../entities/post/types/types.ts';
+import { Comment } from '../../entities/comment/ui/Comment.tsx';
 
 type CommentListProps = {
   post: PostProps;
@@ -10,19 +11,7 @@ export function CommentList({ post }: CommentListProps) {
     <ul className={styles.commentsBlock}>
       {post.comments.map((comment) => (
         <li className={styles.commentsBlock__comment} key={comment.id}>
-          <p className={styles.commentsBlock__comment_text}>
-            {comment.commentText}
-          </p>
-          <div className={styles.commentsBlock__comment_other}>
-            <p>{comment.author}</p>
-            <p>
-              {comment.date.toLocaleDateString('en-US', {
-                year: 'numeric',
-                month: 'long',
-                day: 'numeric',
-              })}
-            </p>
-          </div>
+          <Comment comment={comment} />
         </li>
       ))}
     </ul>
